Change to a LanguageConsole to restore command history.
It's a bit of a kludge: the language is 'Plain Text'. But this seems like the path of least resistance. No autocompletion provided: that's a much bigger job.

Requires IDEA 11 (Nika), so bumping the version to 1.4.0-SNAPSHOT and branching
until I release 1.3.2 from master.

Update the release notes accordingly.

Fixes #42, References #33